made seperate component for team & style fixed
